requirements.txt
django==5.0.4
djangorestframework==3.15.1
poetry==1.8.2
psycopg2==2.9.9
drf-spectacular==0.27.0
requests==2.31.0
django-silk==5.1.0
coverage==7.4.4
celery[redis]==5.4.0
flower==2.0.1